requirements.txt
scipy
opencv-python
numpy
sklearn
easydict
scikit-image
bistiming
eyewitness==1.0.13
dchunk==0.0.1
line-bot-sdk
Pillow
imutils
keyboard
argparse